Add touch device detection and update tag components
