+++ examples/style.css
@@ -30,3 +30,18 @@
     dark:text-white
     [box-shadow:0_0_0_1px_var(--border-color),var(--shadow-size)_var(--shadow-color)]
 }
+
+.input {
+  @apply
+    bg-white
+    hover:bg-gray-50
+    dark:bg-gray-900
+    dark:hover:bg-gray-950
+    outline-2
+    outline-offset-[-1px]
+    outline-blue-600
+    focus-within:outline
+    cursor-text
+    [box-shadow:inset_0_0_0_1px_rgba(0_0_0/0.15),inset_0_2px_5px_0_rgba(0_0_0/0.08)]
+    dark:[box-shadow:inset_0_0_0_1px_rgba(255_255_255/0.15),inset_0_-1px_0_0_rgba(255_255_255/0.05),inset_0_2px_5px_0_rgba(0_0_0/0.15)]
+}
