+++ packages/ariakit-react-core/src/combobox/combobox.tsx
@@ -379,8 +379,7 @@ export const useCombobox = createHook<TagName, ComboboxOptions>(
       onChangeProp?.(event);
       if (event.defaultPrevented) return;
       if (!store) return;
-      const currentTarget = event.currentTarget;
-      const { value, selectionStart, selectionEnd } = currentTarget;
+      const { value, selectionStart } = event.currentTarget;
       const nativeEvent = event.nativeEvent;
       canAutoSelectRef.current = true;
       if (isInputEvent(nativeEvent)) {
@@ -399,17 +398,6 @@ export const useCombobox = createHook<TagName, ComboboxOptions>(
       if (setValueOnChangeProp(event)) {
         const isSameValue = value === store.getState().value;
         store.setValue(value);
-        // flushSync(() => store?.setValue(value));
-        // When the value is not set synchronously, the selection range may be
-        // lost. Even setting the value with flushSync above, we still need to
-        // fix the selection range because React's useSyncExternalStore updates
-        // in a microtask. An alternative fix would be removing the flushSync
-        // call above and calling setSelectionRange in a queueMicrotask
-        // callback, but I think flushSync is a safer approach. See
-        // combobox-group "keep caret position when typing" test.
-        queueMicrotask(() => {
-          setSelectionRange(currentTarget, selectionStart, selectionEnd);
-        });
         if (inline && autoSelect && isSameValue) {
           // The store.setValue(event.target.value) above may not trigger a
           // state update. For example, say the first item starts with "t". The
