+++ packages/ariakit-react-core/src/tag/tag.tsx
@@ -17,10 +17,9 @@ import {
   memo,
 } from "../utils/system.js";
 import type { Props } from "../utils/types.js";
-import { VisuallyHidden } from "../visually-hidden/visually-hidden.js";
-import type { VisuallyHiddenOptions } from "../visually-hidden/visually-hidden.js";
 import { TagValueContext, useTagContext } from "./tag-context.jsx";
 import type { TagStore } from "./tag-store.js";
+import { useTouchDevice } from "./utils.js";
 
 const TagName = "div" satisfies ElementType;
 type TagName = typeof TagName;
@@ -40,7 +39,6 @@ export const useTag = createHook<TagName, TagOptions>(function useTag({
   value,
   getItem: getItemProp,
   removeOnKeyPress = true,
-  hiddenRemove = true,
   ...props
 }) {
   const context = useTagContext();
@@ -53,6 +51,7 @@ export const useTag = createHook<TagName, TagOptions>(function useTag({
   );
 
   const id = useId(props.id);
+  const touchDevice = useTouchDevice();
 
   const getItem = useCallback<NonNullable<CompositeItemOptions["getItem"]>>(
     (item) => {
@@ -68,14 +67,6 @@ export const useTag = createHook<TagName, TagOptions>(function useTag({
   const onKeyDownProp = props.onKeyDown;
   const removeOnKeyPressProp = useBooleanEvent(removeOnKeyPress);
 
-  const removeValue = useEvent((moveToPrevious = false) => {
-    store.removeValue(value);
-    const nextId = moveToPrevious
-      ? store.previous() || store.next()
-      : store.next();
-    store.move(nextId);
-  });
-
   const onKeyDown = useEvent((event: KeyboardEvent<HTMLType>) => {
     onKeyDownProp?.(event);
     if (event.defaultPrevented) return;
@@ -86,7 +77,11 @@ export const useTag = createHook<TagName, TagOptions>(function useTag({
 
     if (isRemoveKey && removeOnKeyPressProp(event)) {
       event.preventDefault();
-      removeValue(isBackspace);
+      store.removeValue(value);
+      const nextId = isBackspace
+        ? store.previous() || store.next()
+        : store.next();
+      store.move(nextId);
     }
 
     const isPrintableKey =
@@ -107,49 +102,18 @@ export const useTag = createHook<TagName, TagOptions>(function useTag({
     [value],
   );
 
-  const removeId = useId();
-
-  // Screen readers don't focus on elements within an option role, and mobile
-  // screen reader users can't hit Backspace while focused on the option. Thus,
-  // we include a visually hidden button with a button role adjacent to the
-  // option. Even if it's not focusable, the button role enables screen readers
-  // to move the virtual cursor and interact with it.
-  props = useWrapElement(
-    props,
-    (element) => {
-      if (!hiddenRemove) return element;
-      const render =
-        typeof hiddenRemove !== "boolean" ? hiddenRemove : <button />;
-      return (
-        <>
-          {element}
-          <VisuallyHidden
-            id={removeId}
-            tabIndex={-1}
-            aria-labelledby={`${removeId} ${id}`}
-            render={render}
-            onClick={() => removeValue()}
-          >
-            Remove
-          </VisuallyHidden>
-        </>
-      );
-    },
-    [removeId, hiddenRemove, removeValue],
-  );
-
   props = {
     id,
-    role: "option",
+    role: !touchDevice ? "option" : "listitem",
     children: value,
     ...props,
+    onKeyDown,
   };
 
   props = useCompositeItem<TagName>({
     store,
     ...props,
     getItem,
-    onKeyDown,
   });
 
   return props;
@@ -215,33 +179,6 @@ export interface TagOptions<T extends ElementType = TagName>
    * @default true
    */
   removeOnKeyPress?: BooleanOrCallback<KeyboardEvent<HTMLType>>;
-  /**
-   * By default, a visually hidden remove button is rendered next to the tag
-   * for mobile screen reader accessibility, where users can't press `Backspace`
-   * while focusing on tags. If set to `false`, this element won't be rendered.
-   *
-   * The accessible label of the remove element is a combination of its own
-   * content, which defaults to the English word "Remove", and the tag's
-   * content. You can customize this by passing an element with a different
-   * content or `aria-label` attribute. When using `aria-label`, you should
-   * manually include the tag's content for context.
-   * @default true
-   * @example
-   * You can pass an element with a different content to customize the "Remove"
-   * label:
-   * ```jsx
-   * // Delete element's label: "Eliminar Manzana"
-   * <Tag value="Manzana" hiddenRemove={<button>Eliminar</button>} />
-   * ```
-   * @example
-   * You can pass a custom `aria-label` attribute to override the entire label.
-   * For example, for languages where the verb doesn't come before the noun:
-   * ```jsx
-   * // Delete element's label: "사과 삭제"
-   * <Tag value="사과" hiddenRemove={<button aria-label="사과 삭제" />} />
-   * ```
-   */
-  hiddenRemove?: boolean | VisuallyHiddenOptions["render"];
 }
 
 export type TagProps<T extends ElementType = TagName> = Props<T, TagOptions<T>>;
